requirements.txt
# Core quantum ML
pennylane==0.38.0
pennylane-lightning==0.38.0

# Deep learning
torch==2.3.0
torchvision==0.18.0

# Classical ML baselines
scikit-learn==1.5.0
scikit-optimize==0.10.1

# Numerics & data
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2

# Experiment tracking
mlflow==2.13.2

# Visualisation
matplotlib==3.9.0
seaborn==0.13.2

# Config & utilities
pyyaml==6.0.1
tqdm==4.66.4

# Development
pytest==8.2.2
pytest-cov==5.0.0
flake8==7.0.0
black==24.4.2
